requirements.txt
# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Segurança
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0

# PDF
reportlab==4.0.7
PyPDF2==3.0.1
Pillow>=10.1.0

# Validação de Dados
marshmallow==3.20.1

# Variáveis de Ambiente
python-dotenv==1.0.0

# CORS (opcional, se precisar de API externa)
Flask-CORS==4.0.0

# Testes
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilitários
python-dateutil==2.8.2

